requirements.txt
Django==1.4
armstrong.core.arm-layout==1.1.1

lxml==2.3.4
requests==0.13.1
